#78-댓글 수정 추가

+++ src/main/java/com/goorp/backend/domain/comment/CommentService.java
@@ -1,17 +1,18 @@
 package com.goorp.backend.domain.comment;
 
-import com.goorp.backend.domain.member.Member;
-import com.goorp.backend.domain.post.Post;
-import com.goorp.backend.domain.comment.model.CommentRequestDto;
-import com.goorp.backend.domain.comment.model.CommentResponseDto;
 import com.goorp.backend.api.exception.CommentException;
 import com.goorp.backend.api.exception.ErrorCode;
+import com.goorp.backend.domain.comment.model.CommentRequestDto;
+import com.goorp.backend.domain.comment.model.CommentResponseDto;
+import com.goorp.backend.domain.comment.model.CommentUpdateDto;
+import com.goorp.backend.domain.member.Member;
 import com.goorp.backend.domain.member.MemberRepository;
+import com.goorp.backend.domain.post.Post;
 import com.goorp.backend.domain.post.PostRepository;
+import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import java.util.List;
 @Service
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
@@ -24,12 +25,8 @@ public class CommentService {
     // CREATE
     @Transactional
     public CommentResponseDto createComment(Long postId, CommentRequestDto request, String accountId) {
-        Post post = postRepository.findById(postId)
-            .orElseThrow(() -> new CommentException(ErrorCode.ID_NOT_FOUND, "postId 가 없습니다."));
-
-        Member member = memberRepository.findByAccountId(accountId)
-            .orElseThrow(
-                () -> new CommentException(ErrorCode.NOT_FOUND_MEMBER,"멤버 ID를 찾을 수 없습니다: " + accountId));
+        Post post = finPostByPostId(postId);
+        Member member = findMemberByMemberId(accountId);
 
         Comment parentComment = null;
         if (request.getParentCommentId() != null) {
@@ -55,13 +52,44 @@ public class CommentService {
             .toList();
     }
 
+    // UPDATE
+    public CommentResponseDto updateComment(Long commentId, CommentUpdateDto commentUpdateDto, Long memberId) {
+        // 검증하고 댓글 내용 업데이트
+        Comment comment = findCommentByCommentId(commentId);
+        validateCommentAuthor(memberId, comment);
+        comment.updateComment(commentUpdateDto.getContent());
+        return CommentResponseDto.of(comment);
+    }
+
     // DELETE
     @Transactional
-    public void deleteComment(Long postId, Long commentId) {
-        postRepository.findById(postId)
-            .orElseThrow(() -> new CommentException(ErrorCode.ID_NOT_FOUND, "postId 가 없습니다."));
-        commentRepository.findById(commentId)
-            .orElseThrow(() -> new CommentException(ErrorCode.ID_NOT_FOUND, "commentId 가 없습니다."));
+    public void deleteComment(Long postId, Long commentId, Long memberId) {
+        finPostByPostId(postId);
+        Comment findComment = findCommentByCommentId(commentId);
+        validateCommentAuthor(memberId, findComment);
+
         commentRepository.deleteById(commentId);
+    }
+
+    private static void validateCommentAuthor(Long memberId, Comment comment) {
+        if (!comment.validation(memberId)) {
+            throw new CommentException(ErrorCode.UNAUTHORIZED, "댓글 작성자가 아닙니다.");
+        }
+    }
+
+    private Comment findCommentByCommentId(Long commentId) {
+        return commentRepository.findById(commentId)
+            .orElseThrow(() -> new CommentException(ErrorCode.ID_NOT_FOUND, "commentId 가 없습니다."));
+    }
+
+    private Member findMemberByMemberId(String accountId) {
+        return memberRepository.findByAccountId(accountId)
+            .orElseThrow(
+                () -> new CommentException(ErrorCode.NOT_FOUND_MEMBER,"멤버 ID를 찾을 수 없습니다: " + accountId));
+    }
+
+    private Post finPostByPostId(Long postId) {
+        return postRepository.findById(postId)
+            .orElseThrow(() -> new CommentException(ErrorCode.ID_NOT_FOUND, "postId 가 없습니다."));
     }
 }
